feat: 优化核心部分功能

+++ src/Traits/DB/BaseOperate.php
@@ -3,7 +3,7 @@
 // +----------------------------------------------------------------------
 // | CatchAdmin [Just Like ～ ]
 // +----------------------------------------------------------------------
-// | Copyright (c) 2017~2021 https://catchadmin.com All rights reserved.
+// | Copyright (c) 2017~2021 https://catchadmin.vip All rights reserved.
 // +----------------------------------------------------------------------
 // | Licensed ( https://github.com/JaguarJack/catchadmin-laravel/blob/master/LICENSE.md )
 // +----------------------------------------------------------------------
@@ -16,11 +16,11 @@ namespace Catch\Traits\DB;
 
 use Catch\Enums\Status;
 use Catch\Exceptions\FailedException;
+use Catch\Facade\Admin;
 use Closure;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Request;
 
@@ -29,22 +29,34 @@ use Illuminate\Support\Facades\Request;
  */
 trait BaseOperate
 {
-    use WithEvents, WithRelations;
+    use WithEvents;
+    use WithRelations;
+    use WithSearch;
 
-    /**
-     * @return mixed
-     */
     public function getList(): mixed
     {
         $fields = property_exists($this, 'fields') ? $this->fields : ['*'];
 
-        $builder = static::select($fields)
-                    ->creator()
-                    ->quickSearch();
+        // 字段访问，获取可读字段
+        if ($this->columnAccess) {
+            $fields = $this->readable($fields);
+        }
+
+        $builder = static::select($fields);
+
+        // 如果有创建人字段
+        if (in_array($this->getCreatorIdColumn(), $this->getFillable())) {
+            $builder = $builder->creator();
+        }
+
+        // 快速搜索
+        if (! empty($this->searchable)) {
+            $builder = $builder->quickSearch(callback: $this->quickSearchCallback);
+        }
 
         // 数据权限
         if ($this->dataRange) {
-           $builder = $builder->dataRange();
+            $builder = $builder->dataRange();
         }
 
         // before list
@@ -54,36 +66,35 @@ trait BaseOperate
 
         // 排序
         if ($this->sortField && in_array($this->sortField, $this->getFillable())) {
-            $builder = $builder->orderBy($this->aliasField($this->sortField), $this->sortDesc ? 'desc' : 'asc');
+            $builder = $builder->orderBy($this->aliasField($this->sortField), $this->getDefaultSortOrder());
         }
 
         // 动态排序
-        $dynamicSortField = Request::get('sortField');
-        if ($dynamicSortField && $dynamicSortField <> $this->sortField) {
-            $builder = $builder->orderBy($this->aliasField($dynamicSortField),  Request::get('order', 'asc'));
+        $dynamicSortField = Request::get($this->dynamicQuerySortField);
+        if ($dynamicSortField && $dynamicSortField != $this->sortField) {
+            $builder = $builder->orderBy($this->aliasField($dynamicSortField), Request::get($this->dynamicQuerySortOrder, 'asc'));
         }
         $builder = $builder->orderByDesc($this->aliasField($this->getKeyName()));
 
+        $limit = Request::get('limit', $this->perPage);
+        // 如果使用回收站
+        if ($this->isUseTrashed()) {
+            return $builder->onlyTrashed()->paginate($limit);
+        }
+
         // 分页
         if ($this->isPaginate) {
-            return $builder->paginate(Request::get('limit', $this->perPage));
+            return $builder->paginate($limit);
         }
 
-        $data = $builder->get();
-        // if set as tree, it will show tree data
-        if ($this->asTree) {
-            return $data->toTree();
-        }
-
-        return $data;
+        // 如果设置 asTree 属性为 true，将会返回树形结构
+        return $builder->get()->when($this->asTree, function ($collection) {
+            return $collection->toTree(id: $this->getKeyName(), pidField: $this->getParentIdColumn());
+        });
     }
-
 
     /**
      * save
-     *
-     * @param array $data
-     * @return mixed
      */
     public function storeBy(array $data): mixed
     {
@@ -101,7 +112,6 @@ trait BaseOperate
     /**
      * create
      *
-     * @param array $data
      * @return false|mixed
      */
     public function createBy(array $data): mixed
@@ -117,16 +127,12 @@ trait BaseOperate
 
     /**
      * update
-     *
-     * @param $id
-     * @param array  $data
-     * @return mixed
      */
     public function updateBy($id, array $data): mixed
     {
         $model = $this->where($this->getKeyName(), $id)->first();
 
-        $updated = $model->fill($this->filterData($data))->save();
+        $updated = $model->fill($this->filterData($data, true))->save();
 
         if ($updated) {
             $this->updateRelations($this->find($id), $data);
@@ -136,12 +142,41 @@ trait BaseOperate
     }
 
     /**
-     * filter data/ remove null && empty string
-     *
-     * @param array $data
-     * @return array
+     * @param  array<int|string>  $condition
+     * @param  array<string, array<int|string>>  $data
      */
-    protected function filterData(array $data): array
+    public function batchUpdate(string $field, array $condition, array $data): bool
+    {
+        try {
+            $batchSQL = 'UPDATE `'.withTablePrefix($this->getTable()).'` SET ';
+
+            foreach ($data as $key => $values) {
+                $batchSQL .= sprintf('`%s` = CASE ', $key);
+                if (count($condition) != count($values)) {
+                    continue;
+                }
+
+                foreach ($values as $index => $value) {
+                    $batchSQL .= sprintf('WHEN %s = %s THEN "%s" ', $field, $condition[$index], $value);
+                }
+
+                $batchSQL .= 'ELSE '.$key.' END, ';
+            }
+
+            $where = ' WHERE '.$field.' IN ('.implode(',', $condition).')';
+
+            $batchSQL = trim($batchSQL, ', ').$where;
+
+            return DB::statement($batchSQL);
+        } catch (\Exception|\Throwable $exception) {
+            throw new FailedException('批量更新报错: '.$exception->getMessage());
+        }
+    }
+
+    /**
+     * filter data/ remove null && empty string
+     */
+    protected function filterData(array $data, $isUpdate = false): array
     {
         // 表单保存的数据集合
         $fillable = array_unique(array_merge($this->getFillable(), $this->getForm()));
@@ -154,31 +189,64 @@ trait BaseOperate
             if (! empty($fillable) && ! in_array($k, $fillable)) {
                 unset($data[$k]);
             }
+        }
 
-            if (in_array($k, [$this->getUpdatedAtColumn(), $this->getCreatedAtColumn()])) {
-                unset($data[$k]);
+        // 如果设置了可写权限
+        if ($this->columnAccess) {
+            // 获取可写入的字段
+            $keys = $this->writable(array_keys($data));
+            foreach ($data as $key => $value) {
+                // 删除 data 中不在可写入的字段
+                if (! in_array($key, $keys)) {
+                    unset($data[$key]);
+                }
             }
         }
 
-        if ($this->isFillCreatorId && in_array($this->getCreatorIdColumn(), $this->getFillable())) {
-            $data['creator_id'] = Auth::guard(getGuardName())->id();
+        // 写入创建时间和更新时间
+        if (! $this->timestamps) {
+            $createdAtColumn = $this->getCreatedAtColumn();
+            if (! $isUpdate && $createdAtColumn) {
+                $data[$createdAtColumn] = time();
+            }
+
+            // 如果是更新，删除 created_at 字段
+            if ($isUpdate && isset($data[$createdAtColumn])) {
+                unset($data[$createdAtColumn]);
+            }
+            // 更新时间字段
+            if ($updatedAtColumn = $this->getUpdatedAtColumn()) {
+                $data[$updatedAtColumn] = time();
+            }
+        }
+
+        // 创建人
+        $creatorColumn = $this->getCreatorIdColumn();
+        if ($this->isFillCreatorId && in_array($creatorColumn, $this->getFillable())) {
+            $creatorId = $data[$creatorColumn] ?? 0;
+            if (! $creatorId && $id = Admin::id()) {
+                $data[$creatorColumn] = $id;
+            }
         }
 
         return $data;
     }
 
-
     /**
      * get first by ID
      *
-     * @param $value
-     * @param null $field
-     * @param string[] $columns
+     * @param  $value
+     * @param  null  $field
+     * @param  string[]  $columns
      * @return ?Model
      */
     public function firstBy($value, $field = null, array $columns = ['*']): ?Model
     {
         $field = $field ?: $this->getKeyName();
+
+        if ($this->columnAccess) {
+            $columns = $this->readable($columns);
+        }
 
         $model = static::where($field, $value)->first($columns);
 
@@ -191,12 +259,8 @@ trait BaseOperate
 
     /**
      * delete model
-     *
-     * @param $id
-     * @param bool $force
-     * @return bool|null
      */
-    public function deleteBy($id, bool $force = false): ?bool
+    public function deleteBy($id, bool $force = false, bool $softForce = false): ?bool
     {
         /* @var Model $model */
         $model = static::find($id);
@@ -213,7 +277,7 @@ trait BaseOperate
             $deleted = $model->delete();
         }
 
-        if ($deleted) {
+        if ($deleted && ! $softForce) {
             $this->deleteRelations($model);
         }
 
@@ -221,24 +285,37 @@ trait BaseOperate
     }
 
     /**
+     * 删除软删除数据
+     */
+    public function deleteTrash($id): mixed
+    {
+        return static::onlyTrashed()->find($id)->forceDelete();
+    }
+
+    /**
      * 批量删除
      *
-     * @param array|string $ids
-     * @param bool $force
-     * @param Closure|null $callback
      * @return true
+     * @throws \Throwable
      */
-    public function deletesBy(array|string $ids, bool $force = false, Closure $callback = null): bool
+    public function deletesBy(array|string $ids, bool $force = false, ?Closure $callback = null): bool
     {
         if (is_string($ids)) {
             $ids = explode(',', $ids);
         }
 
-        DB::transaction(function () use ($ids, $force, $callback){
-            foreach ($ids as $id) {
-                $this->deleteBy($id, $force);
+        DB::transaction(function () use ($ids, $force, $callback) {
+            // 删除软删除数据
+            if ($this->isUseTrashed()) {
+                foreach ($ids as $id) {
+                    $this->deleteTrash($id);
+                }
+            } else {
+                // 软删除
+                foreach ($ids as $id) {
+                    $this->deleteBy($id, $force);
+                }
             }
-
             if ($callback) {
                 $callback($ids);
             }
@@ -248,21 +325,41 @@ trait BaseOperate
     }
 
     /**
+     * 恢复
+     */
+    public function restoreBy(array|string $ids): bool|int
+    {
+        if (is_string($ids)) {
+            $ids = explode(',', $ids);
+        }
+
+        if (count($ids) === 1) {
+            $model = $this->onlyTrashed()->find($ids[0]);
+
+            return $model->restore();
+        }
+
+        $this->whereIn($this->getKeyName(), $ids)
+            ->onlyTrashed()
+            ->update([
+                $this->getDeletedAtColumn() => 0,
+            ]);
+
+        return true;
+    }
+
+    /**
      * disable or enable
-     *
-     * @param $id
-     * @param string $field
-     * @return bool
      */
     public function toggleBy($id, string $field = 'status'): bool
     {
         $model = $this->firstBy($id);
 
-        $status = $model->getAttribute($field) ==  Status::Enable->value() ? Status::Disable->value() : Status::Enable->value();
+        $status = $model->getAttribute($field) == Status::Enable->value() ? Status::Disable->value() : Status::Enable->value();
 
         $model->setAttribute($field, $status);
 
-        if ($model->save() && in_array($this->getParentIdColumn(), $this->getFillable())) {
+        if ($model->save() && in_array($this->getParentIdColumn(), $this->getFillable()) && in_array($field, $this->syncParentFields)) {
             $this->updateChildren($id, $field, $model->getAttribute($field));
         }
 
@@ -270,10 +367,8 @@ trait BaseOperate
     }
 
     /**
-     *
-     * @param array|string $ids
-     * @param string $field
      * @return true
+     * @throws \Throwable
      */
     public function togglesBy(array|string $ids, string $field = 'status'): bool
     {
@@ -281,7 +376,7 @@ trait BaseOperate
             $ids = explode(',', $ids);
         }
 
-        DB::transaction(function () use ($ids, $field){
+        DB::transaction(function () use ($ids, $field) {
             foreach ($ids as $id) {
                 $this->toggleBy($id, $field);
             }
@@ -290,13 +385,11 @@ trait BaseOperate
         return true;
     }
 
-
     /**
      * 递归处理
      *
-     * @param int|array $parentId
-     * @param string $field
-     * @param int $value
+     * @param  int|array  $parentId
+     * @param  int  $value
      */
     public function updateChildren(mixed $parentId, string $field, mixed $value): void
     {
@@ -308,7 +401,7 @@ trait BaseOperate
 
         if ($childrenId->count()) {
             if ($this->whereIn($this->getParentIdColumn(), $parentId)->update([
-                $field => $value
+                $field => $value,
             ])) {
                 $this->updateChildren($childrenId, $field, $value);
             }
@@ -317,9 +410,6 @@ trait BaseOperate
 
     /**
      * alias field
-     *
-     * @param string|array $fields
-     * @return string|array
      */
     public function aliasField(string|array $fields): string|array
     {
@@ -336,11 +426,8 @@ trait BaseOperate
         return $fields;
     }
 
-
     /**
      * get updated at column
-     *
-     * @return string|null
      */
     public function getUpdatedAtColumn(): ?string
     {
@@ -353,10 +440,15 @@ trait BaseOperate
         return $updatedAtColumn;
     }
 
+    protected function isUseTrashed(): bool
+    {
+        $trashed = Request::get('trashed');
+
+        return $trashed && in_array($this->getDeletedAtColumn(), $this->getFillable());
+    }
+
     /**
      * get created at column
-     *
-     * @return string|null
      */
     public function getCreatedAtColumn(): ?string
     {
@@ -369,29 +461,22 @@ trait BaseOperate
         return $createdAtColumn;
     }
 
-    /**
-     *
-     * @return string
-     */
     public function getCreatorIdColumn(): string
     {
         return 'creator_id';
     }
 
     /**
-     *
      * @return $this
      */
     protected function setCreatorId(): static
     {
-        $this->setAttribute($this->getCreatorIdColumn(), Auth::guard(getGuardName())->id());
+        $this->setAttribute($this->getCreatorIdColumn(), Admin::id());
 
         return $this;
     }
 
     /**
-     *
-     * @param string $parentId
      * @return $this
      */
     public function setParentIdColumn(string $parentId): static
@@ -402,8 +487,6 @@ trait BaseOperate
     }
 
     /**
-     *
-     * @param string $sortField
      * @return $this
      */
     protected function setSortField(string $sortField): static
@@ -414,10 +497,9 @@ trait BaseOperate
     }
 
     /**
-     *
      * @return $this
      */
-    protected function setPaginate(bool $isPaginate = true): static
+    public function setPaginate(bool $isPaginate = true): static
     {
         $this->isPaginate = $isPaginate;
 
@@ -438,9 +520,6 @@ trait BaseOperate
         return $this;
     }
 
-    /**
-     * @return array
-     */
     public function getForm(): array
     {
         if (property_exists($this, 'form') && ! empty($this->form)) {
@@ -450,20 +529,23 @@ trait BaseOperate
         return [];
     }
 
+    protected function getDefaultSortOrder(): string
+    {
+        if (property_exists($this, 'sortDesc')) {
+            return $this->sortDesc ? 'desc' : 'asc';
+        }
+
+        return 'desc';
+    }
+
     /**
      * get parent id
-     *
-     * @return string
      */
     public function getParentIdColumn(): string
     {
         return $this->parentIdColumn;
     }
 
-    /**
-     *
-     * @return array
-     */
     public function getFormRelations(): array
     {
         if (property_exists($this, 'formRelations') && ! empty($this->form)) {
@@ -476,7 +558,6 @@ trait BaseOperate
     /**
      * set data range
      *
-     * @param bool $use
      * @return $this
      */
     public function setDataRange(bool $use = true): static
@@ -487,7 +568,18 @@ trait BaseOperate
     }
 
     /**
-     * @param bool $auto
+     * 设置字段访问
+     *
+     * @return $this
+     */
+    public function setColumnAccess(bool $use = true): static
+    {
+        $this->columnAccess = $use;
+
+        return $this;
+    }
+
+    /**
      * @return $this
      */
     public function setAutoNull2EmptyString(bool $auto = true): static
@@ -498,12 +590,51 @@ trait BaseOperate
     }
 
     /**
-     * @param true $is
+     * @return $this
+     */
+    public function asTree(): static
+    {
+        $this->asTree = true;
+
+        return $this;
+    }
+
+    /**
+     * 禁用分页
+     *
+     * @return $this
+     */
+    public function disablePaginate(): static
+    {
+        $this->isPaginate = false;
+
+        return $this;
+    }
+
+    /**
+     * @param  bool  $is
      * @return $this
      */
     public function fillCreatorId(bool $is = true): static
     {
         $this->isFillCreatorId = $is;
+
+        return $this;
+    }
+
+    /**
+     * 设置需要同步的字段
+     *
+     * @param  array|null  $fields
+     * @return $this
+     */
+    public function setSyncParentFields(?array $fields = []): static
+    {
+        if (is_null($fields)) {
+            $this->syncParentFields = [];
+        } else {
+            $this->syncParentFields = array_merge($this->syncParentFields, $fields);
+        }
 
         return $this;
     }
